Merge remote-tracking branch 'origin/caf/caf-wlan/master' into jb_mr1

* origin/caf/caf-wlan/master:
  wlan: Revision 3.2.3.292
  wlan: Routing Action, Assoc, Auth frame to WQ5.
  wlan: Offloads are not working after the roaming.
  wlan: Use numOfChannels to check if channel list is empty
  wlan: Handle EXIT IMPS failure properly.
  wlan: Excessive logging in LIM is causing WD bark.
  wlan: fix for SENDACTIONFRAME channel Number Check
  wlan : Revision 3.2.3.291
  Wlan: Incorrect pointer arithmetic comparison

+++ CORE/WDI/WPAL/src/wlan_qct_pal_packet.c
@@ -479,7 +479,7 @@ WPT_STATIC WPT_INLINE void* itGetOSPktAddrFromDevice( wpt_packet *pPacket )
    }
    else
    {
-     if(skb->data == skb->tail)
+     if((uintptr_t)skb->data == (uintptr_t)skb->tail)
      {
 #ifdef WLAN_BUG_ON_SKB_ERROR
        wpalDevicePanic();
